+++ es/chain/node.js
@@ -16,16 +16,40 @@
  */
 import * as R from 'ramda'
 import Chain from './'
-import Epoch from '../epoch'
+import Node from '../node'
+import Oracle from '../oracle/node'
+import Contract from '../contract/node'
 import formatBalance from '../utils/amount-formatter'
+import TransactionValidator from '../tx/validator'
+
+/**
+ * ChainNode module
+ *
+ * This is the complement to {@link module:@aeternity/aepp-sdk/es/chain}.
+ * @module @aeternity/aepp-sdk/es/chain/node
+ * @export ChainNode
+ * @example import ChainNode from '@aeternity/aepp-sdk/es/chain/node'
+ */
 
 async function sendTransaction (tx, options = {}) {
-  const { waitMined } = R.merge(this.Chain.defaults, options)
+  const { waitMined, verify } = R.merge(this.Chain.defaults, options)
+  // Verify transaction before broadcast
+  if (verify) {
+    const { validation, tx: txObject, txType } = await this.unpackAndVerify(tx)
+    if (validation.length) {
+      throw Object.assign({
+        code: 'TX_VERIFICATION_ERROR',
+        errorData: { validation, tx: txObject, txType },
+        txHash: tx
+      })
+    }
+  }
+
   const { txHash } = await this.api.postTransaction({ tx })
-  return waitMined ? this.poll(txHash, options) : txHash
+  return waitMined ? { ...(await this.poll(txHash, options)), rawTx: tx } : { hash: txHash, rawTx: tx }
 }
 
-async function balance (address, { height, hash, format = true } = {}) {
+async function balance (address, { height, hash, format = false } = {}) {
   const { balance } = await this.api.getAccountByPubkey(address, { height, hash })
   return format ? formatBalance(balance) : balance.toString()
 }
@@ -124,7 +148,23 @@ async function getMicroBlockHeader (hash) {
   return this.api.getMicroBlockHeaderByHash(hash)
 }
 
-const EpochChain = Chain.compose(Epoch, {
+async function txDryRun (txs, accounts, top) {
+  return this.api.dryRunTxs({ txs, accounts, top })
+}
+
+/**
+ * ChainNode Stamp
+ *
+ * This is implementation of {@link module:@aeternity/aepp-sdk/es/chain--Chain}
+ * composed with {@link module:@aeternity/aepp-sdk/es/contract/node--ContractNodeAPI} and {@link module:@aeternity/aepp-sdk/es/oracle/node--OracleNodeAPI}
+ * @function
+ * @alias module:@aeternity/aepp-sdk/es/chain/node
+ * @rtype Stamp
+ * @param {Object} [options={}] - Initializer object
+ * @return {Object} ChainNode instance
+ * @example ChainNode({url: 'https://sdk-testnet.aepps.com/'})
+ */
+const ChainNode = Chain.compose(Node, Oracle, Contract, TransactionValidator, {
   methods: {
     sendTransaction,
     balance,
@@ -139,8 +179,9 @@ const EpochChain = Chain.compose(Epoch, {
     getGeneration,
     getMicroBlockHeader,
     getMicroBlockTransactions,
-    getKeyBlock
+    getKeyBlock,
+    txDryRun
   }
 })
 
-export default EpochChain
+export default ChainNode