+++ es/tx/tx.js
@@ -25,8 +25,9 @@
 import * as R from 'ramda'
 
 import Tx from './'
+import Node from '../node'
+
 import { buildTx, calculateFee } from './builder'
-import Epoch from '../epoch'
 import { TX_TYPE } from './builder/schema'
 import { buildContractId, oracleQueryId } from './builder/helpers'
 
@@ -35,7 +36,7 @@ const CONTRACT_VM_VERSION = 1
 
 async function spendTx ({ senderId, recipientId, amount, payload = '' }) {
   // Calculate fee, get absolute ttl (ttl + height), get account nonce
-  const { fee, ttl, nonce } = await this.prepareTxParams(TX_TYPE.spend, { senderId, ...R.head(arguments) })
+  const { fee, ttl, nonce } = await this.prepareTxParams(TX_TYPE.spend, { senderId, ...R.head(arguments), payload })
 
   // Build transaction using sdk (if nativeMode) or build on `AETERNITY NODE` side
   const { tx } = this.nativeMode
@@ -272,36 +273,39 @@ async function calculateNonce (accountId, nonce) {
 async function prepareTxParams (txType, { senderId, nonce: n, ttl: t, fee: f, gas, absoluteTtl }) {
   const nonce = await (calculateNonce.bind(this)(senderId, n))
   const ttl = await (calculateTtl.bind(this)(t, !absoluteTtl))
-  const fee = calculateFee(f, txType, { gas, params: R.merge(R.last(arguments), { nonce, ttl }) })
+  const fee = calculateFee(f, txType, { showWarning: this.showWarning, gas, params: R.merge(R.last(arguments), { nonce, ttl }) })
   return { fee, ttl, nonce }
 }
 
 /**
  * Transaction Stamp
  *
- * This implementation of {@link module:@aeternity/aepp-sdk/es/tx--Tx} relays
- * the creation of transactions to {@link module:@aeternity/aepp-sdk/es/epoch--Epoch}.
- * This stamp provide ability to create native spend transaction,
- * all other transaction's using Epoch API.
- * As there is no built-in security between Epoch and client communication, it
+ * This is implementation of [Tx](api/tx.md) relays
+ * the creation of transactions to {@link module:@aeternity/aepp-sdk/es/Node}.
+ * This stamp provide ability to create native transaction's,
+ * or transaction's using Node API.
+ * As there is no built-in security between Node and client communication,
+ * creating transaction using {@link module:@aeternity/aepp-sdk/es/Node} API
  * must never be used for production but can be very useful to verify other
  * implementations.
  * @function
  * @alias module:@aeternity/aepp-sdk/es/tx/tx
  * @rtype Stamp
  * @param {Object} [options={}] - Initializer object
- * @param {Object} [options.nativeMode] - Use Native build of transaction's
- * @param {Object} [options.url] - Node url
- * @param {Object} [options.internalUrl] - Node internal url
+ * @param {Boolean} [options.nativeMode=true] options.nativeMode - Use Native build of transaction's
+ * @param {String} options.url - Node url
+ * @param {String} options.internalUrl - Node internal url
  * @return {Object} Transaction instance
  * @example Transaction({url: 'https://sdk-testnet.aepps.com/'})
  */
-const Transaction = Epoch.compose(Tx, {
-  init ({ nativeMode = true }) {
+const Transaction = Node.compose(Tx, {
+  init ({ nativeMode = true, showWarning = false }) {
     this.nativeMode = nativeMode
+    this.showWarning = showWarning
   },
   props: {
-    nativeMode: null
+    nativeMode: null,
+    showWarning: null
   },
   methods: {
     spendTx,